requirements.txt
pandas==1.1.5
crossrefapi==1.5.0
streamlit==0.77.0
requests==2.25.1
wikipedia==1.4.0
Wikidata==0.7.0
numpy==1.19.5
pygbif==0.5.0
